+++ CHANGELOG.md
@@ -5,6 +5,65 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v4.2.1] - 2020-06-20
+- Add custom Macroable trait with Eloquent model support
+
+## [v4.2.0] - 2020-06-19
+- Fix request.guard binding issue when running in console
+
+## [v4.1.1] - 2020-06-19
+- Refactor route parameters to container service binding
+
+## [v4.1.0] - 2020-06-15
+- Add `mergeRules` method to ValidatingTrait
+- Drop PHP 7.2 & 7.3 support from travis
+
+## [v4.0.7] - 2020-05-30
+- Remove default indent size config
+- Add strip_tags validation rule to string fields
+- Specify events queue
+- Add strip_tags validation rule
+
+## [v4.0.6] - 2020-04-12
+- Fix ServiceProvider registerCommands method compatibility
+
+## [v4.0.5] - 2020-04-09
+- Tweak artisan command registration
+- Reverse commit "Convert database int fields into bigInteger"
+- Refactor publish command and allow multiple resource values
+
+## [v4.0.4] - 2020-04-04
+- Fix namespace issue
+
+## [v4.0.3] - 2020-04-04
+- Enforce consistent artisan command tag namespacing
+- Enforce consistent package namespace
+- Drop laravel/helpers usage as it's no longer used
+
+## [v4.0.2] - 2020-03-20
+- Convert into bigInteger database fields
+- Add shortcut -f (force) for artisan publish commands
+- Fix migrations path
+
+## [v4.0.1] - 2020-03-16
+- Update compatibility with Laravel v7.x
+
+## [v4.0.0] - 2020-03-15
+- Upgrade to Laravel v7.1.x & PHP v7.4.x
+
+## [v3.0.4] - 2020-03-13
+- Tweak TravisCI config
+- Add migrations autoload option to the package
+- Tweak service provider `publishesResources`
+- Install felixkiss/uniquewith-validator composer package
+- Ability to enable/disable hashing for each  model in any area without any need to enable/disable the whole area (#26)
+- Remove indirect composer dependency
+- Drop using global helpers
+- Update StyleCI config
+
+## [v3.0.3] - 2019-12-18
+- Tweak ID obscure feature
+
 ## [v3.0.2] - 2019-09-24
 - Add missing laravel/helpers composer package
 
@@ -96,6 +155,20 @@ This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 ## v0.0.1 - 2016-12-20
 - Tag first release
 
+[v4.2.1]: https://github.com/rinvex/laravel-support/compare/v4.2.0...v4.2.1
+[v4.2.0]: https://github.com/rinvex/laravel-support/compare/v4.1.1...v4.2.0
+[v4.1.1]: https://github.com/rinvex/laravel-support/compare/v4.1.0...v4.1.1
+[v4.1.0]: https://github.com/rinvex/laravel-support/compare/v4.0.7...v4.1.0
+[v4.0.7]: https://github.com/rinvex/laravel-support/compare/v4.0.6...v4.0.7
+[v4.0.6]: https://github.com/rinvex/laravel-support/compare/v4.0.5...v4.0.6
+[v4.0.5]: https://github.com/rinvex/laravel-support/compare/v4.0.4...v4.0.5
+[v4.0.4]: https://github.com/rinvex/laravel-support/compare/v4.0.3...v4.0.4
+[v4.0.3]: https://github.com/rinvex/laravel-support/compare/v4.0.2...v4.0.3
+[v4.0.2]: https://github.com/rinvex/laravel-support/compare/v4.0.1...v4.0.2
+[v4.0.1]: https://github.com/rinvex/laravel-support/compare/v4.0.0...v4.0.1
+[v4.0.0]: https://github.com/rinvex/laravel-support/compare/v3.0.4...v4.0.0
+[v3.0.4]: https://github.com/rinvex/laravel-support/compare/v3.0.3...v3.0.4
+[v3.0.3]: https://github.com/rinvex/laravel-support/compare/v3.0.2...v3.0.3
 [v3.0.2]: https://github.com/rinvex/laravel-support/compare/v3.0.1...v3.0.2
 [v3.0.1]: https://github.com/rinvex/laravel-support/compare/v3.0.0...v3.0.1
 [v3.0.0]: https://github.com/rinvex/laravel-support/compare/v2.1.1...v3.0.0